Update download.py
- add version flag
- rename file with correct extension after it has been downloaded
- print download progress
- stronger and faster way of detecting already downloaded files

Previously, the only way to determine if a file had already been
downloaded was to compare the size given by the "content-length"
header to the size of the target file if it exists. Not only is this
a bad way of doing it (it's not impossible that a different file could
have the same size), but you had to depend on and wait for the server
to respond to get the value. This might be fine for most domains, but
Google Drive quickly returns a 403 status after multiple requests in a
short amount of time. There is also the fact that requests to Google
Drive are made one by one (by default). All of this means that it
would take a very long time for the next runs, and you'd almost always
get errors due to the Google Drive 403s. While still being a pretty
viable option, it pales in comparison to using the SHAs provided in
pak_info.json. If a target file exists, its SHA is now compared to the
'control' and if they match, no need to make a request. Combined with
the use of asyncio.to_thread() to compute the SHAs, this significantly
speeds up re-runs of the script (right now, around 30s vs a few minutes
before), and requests are made only when necessary.