Generate readme differently to avoid duplicates
Reolink added duplicate entries for the RLC-410 and RLC-410W in their
list of devices (see be74d8420955840d8ab99a651f5275d030304a02). With the
current code, these entries count as separate devices because we iterate
over each device and then find the matching firmwares.
Instead, iterate over pak_info.json and add each item to a dictionary
whose keys are cleaned up model and hardware version titles that match
this firmware so that, for example, RLC-410 and RLC-410-5MP become the
same device, because they are.
This allows to spot pak_infos that have IDs that don't match anything in
devices.json, even if this shouldn't happen.