Fix for 403
Requests with the word 'python' (case insensitive) in the user agent are
now forbidden. Instead of jumping straight to using the same user agent
a browser would send, let's try this one and see if it will get added to
the reject list. This might give an idea of whether the change is
related to this project.