Add firmware that was missed during an update
On June 28 this firmware was added on the Reolink
website but with the same ID as an existing
one (v3.1.0.2379_23062702), overwriting some of
the information in this repository. This was
probably a mistake on their part but because we
don't record URL changes for an existing firmware,
the new download link was ignored. This adds the
missing data, restores the overwritten info and
assigns a new ID to the original firmware.